development: added classificaiton entities

+++ src/main/java/com/mps/MPSServer/domain/Channel.java
@@ -16,7 +16,7 @@ public class Channel {
     @GeneratedValue(
             strategy = GenerationType.IDENTITY
     )
-    private Long channelId;
+    private Integer channelId;
 
     private String channelName;
